pricing_module/requirements/requirements.txt
django~=4.2.23
djangorestframework~=3.15.2